refactor: page url 변경, path type 예외 처리
