+++ src/main/java/wooteco/subway/admin/domain/Graphs.java
@@ -8,6 +8,9 @@ import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Component;
 
+import wooteco.subway.admin.domain.vo.Graph;
+import wooteco.subway.admin.domain.vo.Path;
+import wooteco.subway.admin.domain.vo.PathType;
 import wooteco.subway.admin.dto.response.PathResponse;
 import wooteco.subway.admin.dto.response.StationResponse;
