+++ src/main/java/wooteco/subway/admin/domain/vo/Graph.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.domain;
+package wooteco.subway.admin.domain.vo;
 
 import java.util.List;
 import java.util.Map;
@@ -7,6 +7,11 @@ import java.util.Objects;
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.WeightedMultigraph;
+
+import wooteco.subway.admin.domain.Line;
+import wooteco.subway.admin.domain.LineStation;
+import wooteco.subway.admin.domain.LineStationEdge;
+import wooteco.subway.admin.domain.Station;
 
 public class Graph {
