+++ src/main/resources/templates/admin/index.html
@@ -27,13 +27,13 @@
     </p>
     <ul class="text-blue-700">
       <li class="inline-block p-1">
-        <a class="underline" href="/admin-stations">역 관리</a>
+        <a class="underline" href="/admin/stations">역 관리</a>
       </li>
       <li class="inline-block p-1">
-        <a class="underline" href="/admin-lines">노선 관리</a>
+        <a class="underline" href="/admin/lines">노선 관리</a>
       </li>
       <li class="inline-block p-1">
-        <a class="underline" href="/admin-edges">구간 관리</a>
+        <a class="underline" href="/admin/edges">구간 관리</a>
       </li>
     </ul>
   </div>
